test for orbital energies
